Tap the screen if we are waiting
So we can accelerate the animation. This commit also change the behavior
of the function. If any of the target images is not found, we are still
in waiting, so tap the screen, sleep 1s and return. If all the images
are located twice, return immediately without waiting for another 1s.