Dedicated function to check stars (max limit break)
Double check the stars

The locations of the craft essence tend to drift a bit after scrolling
down several times, probably due to the inertia effect in the animation.
Double check the stars with a -4 offset in the y direction. Seems to
always produces reasonably high matching score in at least one of the
test.