Better scrolling behavior
First determine if there is more than 2 friends in the list by checking
if there is a scroll bar. When the scroll bar reaches the end of the
list, adjust the offset to check the last two servants/items correctly.

Reset the scrollbar to the top

It seems to happen when there is no filter applied to the friend list,
the scrollbar stays at the same location after switching servant type,
use an extra tap to make sure the scrollbar moves back to the top.