Compensate for small offset after swipe
By location the position of "Suppport" banner we can determine
accurately the location of the support characters. Use a +/-40 window to
make sure it almost always find the landmark target. In the next
swipe, we will compensate the offshoot and reduce the amount of
swipe by 2*offset if offset is < 0; because of the nature of the
inertia there is no need to compensate if offset is > 0, as long as we
do not swipe to fast