Improve stage detection
Threshold the image to remove the background. Since there is a dark
overlay on top of the background image, Threshold at 180 after convert
the image to gray will wipe out all the irrelevant background. There is
a problem related to the resolution issue. When scale up/down for a
different resolution, the exact location of the numbers will change a
little, this seems to cause getIdentityScore call to fail, but findImage
in a +1 window seems to work reasonably well